requirements.txt
# python >= 2.7.10
django==1.9.6
mongoengine==0.10.6
djangorestframework==3.3.3
django-filter==0.13.0
markdown==2.6.6
tld==0.7.6
simplejson==3.8.1